The Center Construction

Let $C$be a fusion category.

AbstractAlgebra.Generic.dim — Method

dim(X::CenterObject)

Return the categorical dimension of X.

AbstractAlgebra.compose — Method

compose(f::CenterMorphism, g::CenterMorphism)

Return the composition g∘f.

AbstractAlgebra.isisomorphic — Method

isisomorphic(X::CenterObject, Y::CenterObject)

Check if X≃Y. Return (true, m) where mis an isomorphism if true, else return (false,nothing).

AbstractAlgebra.kernel — Method

kernel(f::CenterMoprhism)

Return a tuple (K,k) where Kis the kernel object and kis the inclusion.

Base.inv — Method

inv(f::CenterMorphism)

Return the inverse of fif possible.

Base.one — Method

one(C::CenterCategory)

Return the one object of C.

Base.zero — Method

zero(C::CenterCategory)

Return the zero object of C.

Hecke.cokernel — Method

cokernel(f::CenterMoprhism)

Return a tuple (C,c) where Cis the cokernel object and cis the projection.

Hecke.dual — Method

dual(X::CenterObject)

Return the (left) dual object of X.

Hecke.id — Method

id(X::CenterObject)

Return the identity on X.

Hecke.iscentral — Function

iscentral(Z::Object)

Return true if Z is in the categorical center, i.e. there exists a half-braiding on Z.

Hecke.tensor_product — Method

tensor_product(f::CenterMorphism,g::CenterMorphism)

Return the tensor product of f and g.

Hecke.tensor_product — Method

tensor_product(X::CenterObject, Y::CenterObject)

Return the tensor product of X and Y.

LinearAlgebra.tr — Method

tr(f:::CenterMorphism)

Return the categorical trace of f.

TensorCategories.Center — Method

Center(C::Category)

Return the Drinfeld center of C.

TensorCategories.add_simple! — Method

add_simple!(C::CenterCategory, S::CenterObject)

Add the simple object S to the vector of simple objects.

TensorCategories.associator — Method

associator(X::CenterObject, Y::CenterObject, Z::CenterObject)

Return the associator isomorphism (X⊗Y)⊗Z → X⊗(Y⊗Z).

TensorCategories.braiding — Method

braiding(X::CenterObject, Y::CenterObject)

Return the braiding isomorphism X⊗Y → Y⊗X.

TensorCategories.coev — Method

coev(X::CenterObject)

Return the coevaluation morphism 1 → X⊗X∗.

TensorCategories.dsum — Method

dsum(f::CenterMorphism, g::CenterMorphism)

Return the direct sum of f and g.

TensorCategories.dsum — Method

dsum(X::CenterObject, Y::CenterObject)

Return the direct sum object of X and Y.

TensorCategories.ev — Method

ev(X::CenterObject)

Return the evaluation morphism X⊗X → 1.

TensorCategories.half_braiding — Method

half_braiding(Z::CenterObject)

Return a vector with half braiding morphisms Z⊗S → S⊗Z for all simple objects S.

TensorCategories.half_braidings — Method

half_braidings(Z::Object)

Return all objects in the center lying over Z.

TensorCategories.simples — Method

simples(C::CenterCategory)

Return a vector containing the simple objects of C. The list might be incomplete.

TensorCategories.spherical — Method

spherical(X::CenterObject)

Return the spherical structure X → X∗∗ of X.

TensorCategories.zero_morphism — Method

zero_morphism(X::CenterObject, Y::CenterObject)

Return the zero morphism 0:X → Y.