The Grothendieck Group

The Grothendieck group of an abelian category can be computed with

TensorCategories.grothendieck_ring — Function

grothendieck_ring(C::Category)

Return the grothendieck ring of the multiring category $C$.

Missing docstring.

Missing docstring for `````

Partial Grothendieck Rings

Coming Soon`. Check Documenter's build log for details.

@autodocs Modules = [TensorCategories] Pages = ["GrothendieckRing.jl"] ```