Basic Categorical Constructions

Let $\mathcal C$and $\mathcal D$ be categories.

Opposite Category

The opposite category $\mathcal C^{op}$ of $\mathcal C$ has the same objects and morphisms $f\colon Y \to X \in \mathrm{Hom}_{\mathcal C^{op}}(X,Y)$ formally switching domain and codomain.

TensorCategories.opposite_category — Function

opposite_category(C::Category)

Construct the category $Cᵒᵖ$.

TensorCategories.opposite_object — Function

opposite_object(X::Object)

Regard the object $X ∈ C$ as an object in $Cᵒᵖ$.

TensorCategories.opposite_morphism — Function

opposite_morphism(f::Morphism)

Regard the morphism $f ∈ C$ as a morphism in $Cᵒᵖ$.

Product Categories

Given a family of categories $\mathcal C_1,...,\mathcal C_n$ we can form the product category $\mathcal C = \mathcal C_1 \times \cdots \times \mathcal C_n$. Objects and morphism are just families of objects and morphisms. Structures all categories have and are preserved by the product will be available.

TensorCategories.product_category — Function

product_category(C::Category...)

Construct the product category

TensorCategories.product_object — Function

product_object(X::Object...)

Construct the product in the product category

TensorCategories.product_morphism — Function

product_morphism(f::Morphism...)

Construct the product morphism in the product category