Generic Methods

Many constructions for additive, abelian, linear or monoidal categories are entangled. Thus we provide a vast list of methods to compute objects or morphisms using other methods.

Keep in mind that the performance will usually be much better if the following methods are overwritten form custom types.

AbstractAlgebra.image — Method

image(f::Morphism)

Return the image $Im(f)$ of $f:X → Y$ together with a monomorphism $Im(f) ↪ Y$.

Hecke.central_primitive_idempotents — Method

central_primitive_idempotents(H::AbstractHomSpace)

Compute the central primitive idempotents of an endomorphism space $H$.

Hecke.decompose — Method

decompose(X::Object)

Decompose an object $X$ in an abelian category.

Hecke.decompose — Method

decompose(X::Object, S::Vector{Object})

Decompose an object $X$ in a semisimple category into simple objects of $S$.

Hecke.is_simple — Method

is_simple(X::Object)

Check whether $X$ is a simple object.

Nemo.eigenvalues — Method

eigenvalues(f::Morphism)

Compute the eigenvalues of $f$. Return a dictonary with entries λ => ker(f - λid).

TensorCategories.endomorphism_ring — Function

endomorphism_ring(X::Object)

Return the endomorphism ring of $X$ as a matrix algebra.

TensorCategories.express_in_basis — Function

express_in_basis(f::Morphism)

Return a vector of coefficients expressing $f: X → Y$ in the basis o f $\mathrm{Hom}(X,Y)$.

TensorCategories.express_in_basis — Method

express_in_basis(f::Morphism, B::Vector{Morphism})

Return a vector of coefficients expressing $f$in the basis $B$.

TensorCategories.horizontal_direct_sum — Method

function horizontal_direct_sum(f::Morphism, g::Morphism)

Return the sum of $f:X → Z$, $g:Y → Z$ as ``f+g:X⊕Y → Z.

TensorCategories.is_epimorphism — Method

is_epimorphism(f::Morphism)

Check wether $f$is epi.

TensorCategories.is_monomorphism — Method

is_monomoprhism(f::Morphism)

Check whether $f$ mono.

TensorCategories.left_inverse — Method

left_inverse(f::Morphism)

Compute a morphism $g$ such that $g ∘ f = id$. Errors if $f$is not mono.

TensorCategories.right_inverse — Method

right_inverse(f::Morphism)

Compute a morphism $g$ such that $f ∘ g = id$. Errors if $f$ is not epi.

TensorCategories.vertical_direct_sum — Method

function vertical_direct_sum(f::Morphism, g::Morphism)

Return the sum of $f:X → Y$, $g:X → Z$ as ``f+g: X → Y⊕Z.

TensorCategories.zero_morphism — Method

zero_morphism(X::Object, Y::Object)

Compute the zero morphism between $X$and $Y$.

TensorCategories.zero_morphism — Method

zero_morphism(X::Object)

Compute the zero morphism on $X$.

Base.:^ — Method

^(X::Object, n::Integer)

Return the n-fold product object X^n.

Hecke.:⊗ — Method

⊗(f::Morphism, g::Morphism)

Return the tensor product morphism of fand g.

Hecke.:⊗ — Method

⊗(X::Object...)

Return the tensor product object.

TensorCategories.:× — Method

×(X::Object...)

Return the product Object and an array containing the projection morphisms.

TensorCategories.∐ — Method

∐(X::Object...)

Return the coproduct Object and an array containing the injection morphisms.