Representations

We provide a simple abstract type hirarchy for representation categories:

abstract type RepresentationCategory{T} <: TensorCategory{T}

A representation category always requires the field

base_ring::Field

Representations of Finite groups

Let $G$ be a finite group. We consider the category of finite dimensinal $k$-representations of $G$.

GroupRepresentationCategory{T,G} <: RepresentationCategory{T}

Build it with the constructor

JuCat.RepresentationCategory — Method

RepresentationCategory(G::GAPGroup, F::Field)

Category of finite dimensonal group representations of \G.

A group representation is defined by a group homomorphism from $G$ into a finite dimensional vector space $k^n$. These objects are of type

GroupRepresentation{T,G} <: Representation

They are constructed in one of two ways, either by images of generators or by a function

JuCat.Representation — Method

Representation(G::GAPGroup, pre_img::Vector, img::Vector)

Group representation defined by the images of generators of G.

JuCat.Representation — Method

Representation(G::GAPGroup, m::Function)

Group representation defined by m:G -> Mat_n.

where in both cases the images are required to be fitting MatrixElem objects.

Since group reprsentation categories are tensor categories, we again have methods for the important operations