Groethendieck Ring

TensorCategories can compute the Groethendieck ring of a semisimple multitensor category, if provided with methods

# Return the simple objects
simples(C::YourCategory) ::Vector{YourCategoryObject}

# Decompose into simple summands with multiplicity
decompose(X::YourCategoryObject) ::Vector{Tuple{YourCategoryObject, Int}}

# Check for isomorphy
is_isomorphic(X::YourCategoryObject, Y::YourCategoryObject) ::Tuple{Bool, Union{YourMorphism, Nothing}}

Provided these methods exist you can call the function grothendieck_ring on your category.

TensorCategories.grothendieck_ring — Function

grothendieck_ring(C::Category)

Return the grothendieck ring of the multiring category C.

Since this is a generic methods it is by nature not the fastest. Thus if you know more about your category you should specify grothendieck_ring on your type manually.

Example

Take a look at the convolution category:

G = symmetric_group(2)
X = gset(G, [1,2,3])
F,_ = FiniteField(5)
Conv = ConvolutionCategory(X,F)

R,f = grothendieck_ring(Conv)

The S-Matrix

For a braided fusion category with spherical structure and simple objects $X_i$ the $S$-matrix is defined as

\[\begin{aligned} (S)_{ij} = Tr(c_{X_iX_j} \circ c_{X_jX_i}) \end{aligned}\]

Hence you see you need to provide a trace function and a braiding. The braiding

braiding(X::YourCategoryObject, Y::YourCategoryObject) ::YourMorphism

There are two ways to archive the trace: either you provide the trace by yourself

tr(f::YourMorphism) ::YourMorphism

or you provide a spherical structure, duals, evaluation and coevaluation

spherical(X::YourCategoryObject) ::YourMorphism
dual(X::YourCategoryObject) ::YourCategoryObject
ev(X::YourCategoryObject) ::YourMorphism
coev(X::YourCategoryObject) ::YourMorphism

with which the generic left categorical trace will be computed.

LinearAlgebra.tr — Function

tr(f:::CenterCategoryMorphism)

Return the categorical trace of f.

Example

We can compute now the $S$-matrix of for the representations.

G = symmetric_group(3)
F,_ = FiniteField(5)
Rep = RepresentationCategory(G,F)

smatrix(Rep)
[1 1 2; 1 1 2; 2 2 4]