The Category of Finite Dimensional VectorSpaces

The simplest example to provide are the finite dimensional vector spaces over a field.

The basic methods to generate this are

JuCat.VectorSpaces — Type

VectorSpaces{T}(K::S) where T <: FieldElem

The category of finite dimensional vector spaces over K.

The following methods are provided:

JuCat.GradedVectorSpaces — Type

GradedVectorSpaces{T}(K::S) where T <: FieldElem

The category of finite dimensional vector spaces over K.

JuCat.VectorSpaceObject — Type

VectorSpaceObject{T}

An object in the category of finite dimensional vector spaces.

JuCat.dsum — Method

dsum(X::VectorSpaceObject{T,S}...) where {T,S <: FreeModule}

Direct sum space of X... together with the embedding morphisms.

#The Category of Graded Vector Spaces