Coherent sheaves on Finite Sets

Coherent sheaves on a finite set $X$ are characterised by the stalks of the the elements $x \in X$. I.e. a coherent sheaf on $X$ is a collection of vector spaces $V_x$ for each $x\in X$. More generally for a group $G$ and a G-set $X$ we can define consider $G$-equivariant sheaves on $X$. In this case a sheaf is specified by representations of the stabilizers of representatives for the orbits.

Equivariant Coherent Sheaves

We provide the datatype

CoherentSheaf <: Object

The category of equivariant coherent sheafes has type

CohSheaves <: MultiTensorCategory

and can be constructed via

TensorCategories.CohSheaves — Type

CohSheaves(X::GSet,F::Field)

The category of $G$-equivariant coherent sheafes on $X$.

CohSheaves(X, F::Field)

The category of coherent sheafes on $X$.

Morphisms are given by morphisms of representations of the stalks and are of type

CohSheafMorphism{T,G} <: Morphism

TensorCategories.CohSheaves — Method

CohSheaves(X, F::Field)

The category of coherent sheafes on $X$.

TensorCategories.CohSheaves — Method

CohSheaves(X::GSet,F::Field)

The category of $G$-equivariant coherent sheafes on $X$.

AbstractAlgebra.isisomorphic — Method

isisomorphic(X::CohSheaf, Y::CohSheaf)

Check whether $X$and $Y$ are isomorphic and the isomorphism if possible.

Base.one — Method

one(C::CohSheaves)

Return the one object in $C$.

Base.zero — Method

zero(C::CohSheaves)

Return the zero sheaf on the $G$-set.

Hecke.decompose — Method

decompose(X::CohSheaf)

Decompose $X$ into a direct sum of simple objects with multiplicity.

Hecke.tensor_product — Method

tensor_product(X::CohSheaf, Y::CohSheaf)

Return the tensor product of equivariant coherent sheaves.

Hecke.tensor_product — Method

tensor_product(f::CohSheafMorphism, g::CohSheafMorphism)

Return the tensor product of morphisms of equivariant coherent sheaves.

Oscar.issemisimple — Method

issemisimple(C::CohSheaves)

Return whether $C$is semisimple.

TensorCategories.Hom — Method

Hom(X::CohSheaf, Y::CohSheaf)

Return Hom($X,Y$) as a vector space.

TensorCategories.dsum — Function

dsum(X::CohSheaf, Y::CohSheaf, morphisms::Bool = false)

Return the direct sum of sheaves. Return also the inclusion and projection if morphisms = true.

TensorCategories.dsum — Method

dsum(f::CohSheafMorphism, g::CohSheafMorphism)

Return the direct sum of morphisms of sheaves.

TensorCategories.simples — Method

simples(C::CohSheaves)

Return the simple objects of $C$.

TensorCategories.stalks — Method

stalks(X::CohSheaf)

Return the stalks of $X$.

Convolution Category

The objects of this category are again $G$-equivariant coherent sheaves on a finite $G$-set $X\times X$. But we endow them with a different monoidal product.

Let $p_{ij}: X\times X\times X \to X \times X$ be the canonical projections. Then we define the monoidal product of two coherent sheaves $x$ and $y$

\[\begin{aligned} x\otimes y = p_{13}_\ast(p_{12}^\ast(x)\otimes' p_{23}^\ast(y)) \end{aligned}\]

where $\otimes'$is the monoidal product of $Coh(X\times X\times X)$. Similar for morphisms.

Objects in this category are of type

ConvolutionObject <: Object

while the convolution category is of type

ConvolutionCategory <: MultiTensorCategory

and can be constructed by

TensorCategories.ConvolutionCategory — Type

ConvolutionCategory(X::GSet, K::Field)

Return the category of equivariant coherent sheaves on $X$ with convolution product.

ConvolutionCategory(X, K::Field)

Return the category of coherent sheaves on $X$ with convolution product.

Morphisms are just morphisms of coherent sheaves with the new tensor product.

ConvolutionMorphism <: Morphism

TensorCategories.ConvolutionCategory — Method

ConvolutionCategory(X, K::Field)

Return the category of coherent sheaves on $X$ with convolution product.

TensorCategories.ConvolutionCategory — Method

ConvolutionCategory(X::GSet, K::Field)

Return the category of equivariant coherent sheaves on $X$ with convolution product.

AbstractAlgebra.isisomorphic — Method

isisomorphic(X::ConvolutionObject, Y::ConvolutionObject)

Check whether $X$and $Y$are isomorphic. Return the isomorphism if true.

Base.one — Method

one(C::ConvolutionCategory)

Return the one object in Conv($X$).

Base.zero — Method

zero(C::ConvolutionCategory)

Return the zero object in Conv($X$).

Hecke.decompose — Method

decompose(X::ConvolutionObject)

Decompose $X$ into a direct sum of simple objects with multiplicities.

Hecke.tensor_product — Method

tensor_product(f::ConvolutionMorphism, g::ConvolutionMorphism)

Return the convolution product of morphisms of coherent sheaves.

Hecke.tensor_product — Method

tensor_product(X::ConvolutionObject, Y::ConvolutionObject)

Return the convolution product of coherent sheaves.

Oscar.issemisimple — Method

issemisimple(C::ConvolutionCategory)

Check whether $C$ semisimple.

TensorCategories.Hom — Method

Hom(X::ConvolutionObject, Y::ConvolutionObject)

Return Hom($X,Y$) as a vector space.

TensorCategories.dsum — Function

dsum(X::ConvolutionObject, Y::ConvolutionObject, morphisms::Bool = false)

documentation

TensorCategories.dsum — Method

dsum(f::ConvolutionMorphism, g::ConvolutionMorphism)

Return the direct sum of morphisms of coherent sheaves (with convolution product).

TensorCategories.orbit_stabilizers — Method

orbit_stabilizers(C::ConvolutionCategory)

Return the stabilizers of representatives of the orbits.

TensorCategories.simples — Method

simples(C::ConvolutionCategory)

Return a list of simple objects in Conv($X$).

TensorCategories.stalks — Method

stalks(X::ConvolutionObject)

Return the stalks of the sheaf $X$.