Representations

We provide a simple abstract type hirarchy for representation categories:

abstract type RepresentationCategory <:Category

A representation category always requires the field

base_ring::Field

Representations of Finite groups

Let $G$ be a finite group. We consider the category of finite dimensinal $k$-representations of $G$.

GroupRepresentationCategory <: RepresentationCategory

Build it with the constructor

TensorCategories.RepresentationCategory — Method

RepresentationCategory(G::GAPGroup, F::Field)

Category of finite dimensonal group representations of \G.

A group representation is defined by a group homomorphism from $G$ into a finite dimensional vector space $k^n$. These objects are of type

GroupRepresentation <: Representation

They are constructed in one of two ways, either by images of generators or by a function

TensorCategories.Representation — Method

Representation(G::GAPGroup, pre_img::Vector, img::Vector)

Group representation defined by the images of generators of G.

TensorCategories.Representation — Method

Representation(G::GAPGroup, m::Function)

Group representation defined by m:G -> Mat_n.

where in both cases the images are required to be fitting MatrixElem objects.

Since group reprsentation categories are tensor categories, we again have methods for the important operations

TensorCategories.Morphism — Method

Morphism(ρ::GroupRepresentation, τ::GroupRepresentation, m::MatElem; check = true)

Morphism between representations defined by $m$. If check == false equivariancy will not be checked.

AbstractAlgebra.isisomorphic — Method

isisomorphic(σ::GroupRepresentation, τ::GroupRepresentation)

Check whether σ and τ are isomorphic. If true return the isomorphism.

Base.one — Method

one(Rep::GroupRepresentationCategory)

Return the trivial representation.

Base.parent — Method

parent(ρ::GroupRepresentation)

Return the parent representation category of ρ.

Base.zero — Method

zero(Rep::GroupRepresentationCategory)

Return the zero reprensentation.

Hecke.decompose — Method

decompose(σ::GroupRepresentation)

Decompose the representation into a direct sum of simple objects. Return a list of tuples with simple objects and multiplicities.

Hecke.id — Method

id(ρ::GroupRepresentation)

Return the identity on ρ.

Hecke.tensor_product — Method

tensor_product(ρ::GroupRepresentation, τ::GroupRepresentation)

Return the tensor product of representations.

Hecke.tensor_product — Method

tensor_product(f::GroupRepresentationMorphism, g::GroupRepresentationMorphism)

Return the tensor product of morphisms of representations.

Oscar.issemisimple — Method

issemisimple(C::GroupRepresentationCategory)

Return true if C is semisimple else false.

TensorCategories.Hom — Method

Hom(σ::GroupRepresentation, τ::GroupRepresentation)

Return the hom-space of the representations as a vector space.

TensorCategories.dsum — Function

dsum(ρ::GroupRepresentation, τ::GroupRepresentation, morphisms::Bool = false)

Return the direct sum of representations. If morphisms is set true inclusion and projection morphisms are also returned.

TensorCategories.dsum — Method

dsum(f::GroupRepresentationMorphism, g::GroupRepresentationMorphism)

Direct sum of morphisms of representations.

TensorCategories.simples — Method

simples(Rep::GroupRepresentationCategory)

Return a list of the simple objects in Rep.