Fusion Categories from 6j-Symbols

In most literature concerning fusion categories they are characterized by so called $6j$-symbols. Often only those and the fusion rules are provided or of interest. Thus we provide a datatype that allows to get a workable fusion category from provided `6j-symbols.

$6j$-Symbols

Let $\mathcal C$ be a locally finite semisimple multitensor category. Then, if $\{X_i\mid i \in \mathcal I\}$ is a collection of the non-isomorphic simple objects, there is an equivalence of abelian categories

\[F \colon \mathcal C \cong \bigoplus\limits_{i \in \mathcal I} \mathrm{Vec}_k\]

given by

\[X \mapsto \mathrm{Hom}(X_i,X).\]

We define $H_{ij}^k := /mathrm{Hom}(X_k, X_i\otimes X_j)$ to be the multiplicity spaces. Now considering the image of a tensor product $X_i \otimes X_j$ of two simple objects we obtain

\[X_i \otimes X_j \mapsto \bigoplus\limits_{k \in \mathcal I} H_{ij}^k\]

After fixing a natural isomorphism

\[(X_i \otimes X_j) \otimes X_k \cong X_i \otimes (X_j \otimes X_k)\]

we obtain morphisms

\[\bigoplus\limits_{k ∈ I} H_{ij}^k \xrightarrow\]

Hecke.dual — Method

dual(X::SixJObject)

Return the dual object of $X$. An error is thrown if $X$ is not rigid.

TensorCategories.associator — Method

associator(X::SixJObject, Y::SixJObject, Z::SixJObject)

Return the associator isomorphism (X⊗Y)⊗Z → X⊗(Y⊗Z).

TensorCategories.extension_of_scalars — Function

extension_of_scalars(X::SixJObject, K::Field)

Return the object $X$ as an object of the category $C⊗K$.

TensorCategories.extension_of_scalars — Method

extension_of_scalars(C::SixJCategory, K::Field)

Return the category $C⊗K$.

TensorCategories.extension_of_scalars — Method

extension_of_scalars(f::SixJmorphism, K::Field)

Return the category $C⊗K$.