Vector Space Categories

Vector spaces in JuCat are of the abstract type

VectorSpaceObject{T} <: Object

They always need the following fields:

basis::Vector{Any}
parent::Category

The Category of Finite Dimensional VectorSpaces

The simplest example to provide are the finite dimensional vector spaces over a field. This category has type

VectorSpaces{T} <: TensorCategory{T}

and can be constructed like so:

using Oscar #hide
F = FiniteField(5,2)
Vec = VectorSpaces(F)

Objects of this category are of the type

VSObject{T} <: VectorSpaceObject{T}

Every vector space object is defined by a basis and a base field provided by the parent category.

JuCat.VectorSpaceObject — Type

VectorSpaceObject{T}

An object in the category of finite dimensional vector spaces.

JuCat.VectorSpaceObject — Method

VectorSpaceObject(Vec::VectorSpaces, n::Int64)
VectorSpaceObject(K::Field, n::Int)
VectorSpaceObject(Vec::VectorSpaces, basis::Vector)
VectorSpaceObject(K::Field, basis::Vector)

The n-dimensional vector space with basis v1,..,vn (or other specified basis)

The Category of Graded Vector Spaces