Tambara Yamagami Categories

Tambara and Yamagami classified a class of near-group categories of multiplicity one over algebraically closed fields in [2]. Let $A$ be an abelian group. Choose a square root $\tau \in k$ of $\vert A\vert$ and a non-degenerate symmetric bilinear form $\chi \colon A \times A \to k^\times$. Then the Tambara-Yamagimi category $TY(A,\chi,\tau)$ has objects $A \cup \{m\}$ with fusion rules

\[\begin{align*} a ⊗ b = a+b,~~a \otimes m = m \otimes a = m,~~ m \otimes m = \sum\limits_{a \in G} a \end{align*}\]

for $a,b \in G$. The non-trivial associativity constraints are given by

\[\begin{align*} a_{a,m,b} = \chi(a,b)\mathrm{id}_m & a_{m,a,m} = \bigoplus\limits_{b\in A}\chi(a,b)\mathrm{id}_b & a_{m,m,m} = \bigoplus\limits_{a,b \in A} \frac{1}{\tau\chi(a,b)}\mathrm{id}_m. \end{align*}\]

Those categories can be constructed with a generic symmetric bilinear form or with a custom bilinear form and over arbitrary fields.

TensorCategories.tambara_yamagami — Function

tambara_yamagami(A::Group)

Construct $TY(A,τ,χ)$ over ℚ̅ where $τ = √|A|$ and $χ$ is a generic non-degenerate bilinear form.

tambara_yamagami(K::ring, A::Group)

Construct $TY(A,τ,χ)$ over $K$ where $τ = √|A|$ and $χ$ is a generic non-degenerate bilinear form.

tambara_yamagami(K::Ring, A::Group, τ::RingElem)

Construct $TY(A,τ,χ)$ over $K$ where $χ$ is a generic non-degenerate bilinear form.

tambara_yamagami(K::Ring, A::Group, τ::RingElem)

Construct $TY(A,τ,χ)$ over $K$ where $τ = √|A|$.

tambara_yamagami(K::Ring, A::Group, τ::RingElem, χ::BilinearForm)

Construct the Category $TY(A,τ,χ)$.

Tambara-Yamagami categories are implemented as an instance of SixJCategory and hence all functionality follows from there.

Ising Category

The Ising fusion category is a special example of a Tambara-Yamagami category with $A = \mathbb Z_2$.

TensorCategories.Ising — Function

Ising()

Construct the Ising category over $ℚ(√2)$.

Ising(F::Ring)

Construct the Ising category over $F$.

Ising(F::Ring, τ::RingElem)

Construct the Ising category with specific $τ = √2$.

Ising(F::Ring, q::Int)

Construct the braided Ising category over $F$ where q = ±1 defined the braiding defined by ±i.

Ising(F::Ring, τ::RingElem, q::Int)

Construct the Ising fusion category where $τ = √2$ a root and q ∈ {1,-1} specifies the braiding if it exists.

The Haagerup Subfactor

The fusion categories stemming from the Haagerup subfactor are a well known and important example of a fusion category. Details can be found in [3].

In the Morita equivalence class of the Haagerup subfactor lie three categories. We call them $\mathcal H_1,\mathcal H_2$ and $\mathcal H_3$. The first two have multiplicity larger then 1 and are up to now not included. The third has multiplicity 1 and is also known as the Haagerup-Izumi category for $\mathbb Z_3$. It has six simple objects and fusion rules

\[\begin{array}{c||c|c|c|c|c|c} & \mathbb 1 & \alpha & \alpha^\ast & \rho & {}_{\alpha}\rho & {}_{\alpha^\ast}\rho \\ \hline \hline \mathbb 1 & \mathbb 1 & \alpha & \alpha^\ast & \rho & {}_{\alpha}\rho & {}_{\alpha^\ast}\rho \\ \hline \alpha & \alpha & \alpha^\ast & \mathbb 1 & {}_\alpha\rho & {}_{\alpha^\ast}\rho & \rho \\ \hline \alpha^\ast & \alpha^\ast & \mathbb 1 & \alpha & {}_{\alpha^\ast}\rho & \rho & {}{\alpha}\rho \\ \hline \rho & \rho & {}_{\alpha^\ast}\rho & {}_\alpha\rho & \mathbb 1 \oplus \rho \oplus {}_\alpha\rho \oplus {}_{\alpha^\ast}\rho & \alpha \oplus \rho \oplus {}_\alpha\rho \oplus {}_{\alpha^\ast}\rho & \alpha^\ast \oplus \rho \oplus {}_\alpha\rho \oplus {}_{\alpha^\ast}\rho \\ \hline {}_\alpha\rho & {}_\alpha\rho & {}_{\alpha^\ast}\rho & \rho & \alpha \oplus \rho \oplus {}_\alpha\rho \oplus {}_{\alpha^\ast}\rho & \mathbb 1 \oplus \rho \oplus {}_\alpha\rho \oplus {}_{\alpha^\ast}\rho & \alpha^\ast \oplus \rho \oplus {}_\alpha\rho \oplus {}_{\alpha^\ast}\rho \\ \hline {}_{\alpha^\ast}\rho & {}_{\alpha^\ast}\rho & \rho & {}_\alpha\rho & \alpha^\ast \oplus \rho \oplus {}_\alpha\rho \oplus {}_{\alpha^\ast}\rho & \alpha \oplus \rho \oplus {}_\alpha\rho \oplus {}_{\alpha^\ast}\rho & \mathbb 1 \oplus \rho \oplus {}_\alpha\rho \oplus {}_{\alpha^\ast}\rho \end{array}\]

We implement the category $\mathcal H_3$ as an instance of SixJCategory. The other two will follow as soon as we know the proper $6j$-symbols.

TensorCategories.haagerup_H3 — Function

haagerup_H3([p1 = 1, p2 = 2])

Build the Haagerup ℋ₃ subfactor category. The category is build as SixJCategory. The associators are taken from the paper

https://arxiv.org/pdf/1906.01322

where p1,p2 = ±1 are parameters for the different possible sets of associators.

Fusion Categories From Truncated Hecke Categories

TODO: Explanation

TensorCategories.I2 — Function

I2(m::Int)
I2(m::Int, K::Ring)

Creates the categorification of biggest cell in I2(m).

TensorCategories.I2subcategory — Function

I2subcategory(m::Int)
I2subcategory(m::Int, R::Ring)

Creates the fusion subcategory of I2.

Various Other Categories Given by $6J$-Symbols

Here are some more examples to play around.

Categorifications by Vercleyen and Singerland

In [4] they found a huge number of fusion rings and some explicit categorifications that are neither Tambara-Yamagami nor Haagerup-Izumi categories.

FR${}_2^{82}$

They provide 97 different (but maybe equivalent) associators one can access.

TensorCategories.cat_fr_8122 — Function

cat_fr_8122(n::Int64)

Categorification of fusion ring FR8211. n chooses one of 96 possibily equivalent sets of associators.

FR${}_3^{94}$

We have a singel associator for this ring.

TensorCategories.cat_fr_9143 — Function

cat_fr_9143()

Categorification of fusion ring FR9143.