Vector Space Categories

Vector spaces in TensorCategories are of the abstract type

abstract type VectorSpaceObject <: Object end

All objects with vector space structure like hom-spaces are and should be implemented as a subtype of this type. They always need the following fields:

basis::Vector
parent::Category

Finite Dimensional VectorSpaces

The simplest example to provide are the finite dimensional vector spaces over a field. This category has type

VectorSpaces <: TensorCategory

and can be constructed like so:

F,a = FiniteField(5,2)
Vec = VectorSpaces()

Objects of this category are of the type

VSObject <: VectorSpaceObject

Every vector space object is defined by a basis and a base field provided by the parent category.

TensorCategories.VectorSpaceObject — Type

VectorSpaceObject

An object in the category of finite dimensional vector spaces.

TensorCategories.VectorSpaceObject — Method

VectorSpaceObject(Vec::VectorSpaces, n::Int64)
VectorSpaceObject(K::Field, n::Int)
VectorSpaceObject(Vec::VectorSpaces, basis::Vector)
VectorSpaceObject(K::Field, basis::Vector)

The n-dimensional vector space with basis v1,..,vn (or other specified basis)

Morphisms in this Category are defined only by matrices of matching dimensions. They are typed as

VSMorphism <: Morphism

and constructed giving a domain, codomain and matrix element.

Morphism(::VectorSpaceObject, ::VectorSpaceObject, ::MatElem)

Graded Vector Spaces

Very similar we have the category of finite dimensional (twisted) $G$-graded vector spaces for a finite group $G$. We have the type

GradedVectorSpaces <: VectorSpaces

and they are constructed in straightforward manner

G = symmetric_group(6)
VecG = GradedVectorSpaces(G)

Category of G-graded vector spaces over Field of algebraic numbers where G is Sym( [ 1 .. 6 ] )

To add a non-trivial associator (twist) there is another constructor.

TensorCategories.TwistedGradedVectorSpaces — Function

TwistedGradedVectorSpaces(G::GAPGroup, i::Int)

Construct the category of twisted graded vectorspaces with the i-th 3-cocycle.

Graded vector spaces decompose into direct sums of vector spaces for each element in $G$.

GVSObject <: VectorSpaceObject

G = symmetric_group(5)
g,s = gens(G)
V1 = VectorSpaceObject(QQ,5)
V2 = VectorSpaceObject(QQ, [:v, :w])
W = VectorSpaceObject(g => V1, s => V2, g*s => V1⊗V2)

Graded vector space of dimension 17 with grading
PermGroupElem[(1,2,3,4,5), (1,2,3,4,5), (1,2,3,4,5), (1,2,3,4,5), (1,2,3,4,5), (1,2), (1,2), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5)]

Morphisms are implemented analogously by pairs of group elements and vector space objects.

GVSMorphism <: Morphism

The constructor is given by

Missing docstring.

Missing docstring for Morphism(::GVSObject, ::GVSObject,::MatElem) where {G, S <: VectorSpaceMorphism}. Check Documenter's build log for details.

Functionality

(Graded) vector spaces form a fusion category. Thus the methods for direct sums, tensor products, dual, one and zero object are all implemented.

AbstractAlgebra.Generic.dim — Method

dim(V::VectorSpaceObject) = length(V.basis)

Return the vector space dimension of $V$.

AbstractAlgebra.direct_sum — Method

function direct_sum(V::GVSObject, W::GVSObject)

Return the direct sum object $V⊕W$.

AbstractAlgebra.direct_sum — Method

direct_sum(f::VectorSpaceMorphism{T},g::VectorSpaceMorphism{T}) where T

Return the direct sum of morphisms of vector spaces.

AbstractAlgebra.direct_sum — Method

direct_sum(X::VectorSpaceObject{T}, Y::VectorSpaceObject{T}) where {T}

Direct sum of vector spaces together with the embedding morphisms.

AbstractAlgebra.is_isomorphic — Method

function is_isomorphic(V::GVSObject, W::GVSObject)

Check whether $V$and $W$are isomorphic as $G$-graded vector spaces and return an isomorphism in the positive case.

AbstractAlgebra.is_isomorphic — Method

is_isomorphic(V::VSObject, W::VSObject)

Check whether $V$ and $W$are isomorphic. Return the isomorphisms if existent.

AbstractAlgebra.kernel — Method

function kernel(f::GVSMorphism)

Return the graded vector space kernel of $f$.

Base.one — Method

function one(C::GradedVectorSpaces)

Return $k$ as the one dimensional graded vector space.

Base.one — Method

one(Vec::VectorSpaces) = VectorSpaceObject(base_ring(Vec),1)

Return the one-dimensional vector space.

Base.zero — Method

function zero(C::GradedVectorSpaces)

Return the zero diemsnional graded vector space.

Base.zero — Method

zero(Vec::VectorSpaces) = VectorSpaceObject(base_ring(Vec), 0)

Return the zero-dimensional vector space.

Hecke.cokernel — Method

function cokernel(f::GVSMorphism)

Return the graded vector space cokernel of $f$.

Hecke.decompose — Method

function decompose(V::GVSObject)

Return a vector with the simple objects together with their multiplicities $[V:Xi]$.

Hecke.dual — Method

function dual(V::GVSObject)

Return the graded dual vector space of $V$.

Hecke.id — Method

id(X::VectorSpaceObject{T}) where T

Return the identity on the vector space $X$.

Hecke.tensor_product — Method

function tensor_product(V::GVSObject, W::GVSObject)

Return the tensor product $V⊗W$.

Hecke.tensor_product — Method

tensor_product(f::VectorSpaceMorphism, g::VectorSpaceMorphism)

Return the tensor product of vector space morphisms.

Hecke.tensor_product — Method

tensor_product(X::VectorSpaceObject{T}, Y::VectorSpaceObject{T}) where {T,S1,S2}

Return the tensor product of vector spaces.

TensorCategories.Hom — Method

function Hom(V::GVSObject, W::GVSObject)

Return the space of morphisms between graded vector spaces $V$ and $W$.

TensorCategories.Hom — Method

Hom(X::VectorSpaceObject, Y::VectorSpaceObject)

Return the Hom($X,Y$`) as a vector space.

TensorCategories.associator — Method

function associator(U::GVSObject, V::GVSObject, W::GVSObject)

return the associator isomorphism $(U⊗V)⊗W → U⊗(V⊗W)$.

TensorCategories.associator — Method

associator(X::VectorSpaceObject, Y::VectorSpaceObject, Z::VectorSpaceObject)

Return the associator isomorphism a::(X⊗Y)⊗Z -> X⊗(Y⊗Z).

TensorCategories.ev — Method

function ev(V::GVSObject)

Return the evaluation map $V*⊗V → 𝟙$.

TensorCategories.simples — Method

function simples(C::GradedVectorSpaces)

Return a vector containing the simple objects of $C$.