Representations

We provide a simple abstract type hirarchy for representation categories:

abstract type RepresentationCategory{T} <: TensorCategory{T}

A representation category always requires the field

base_ring::Field

Representations of Finite groups

Let $G$ be a finite group. We consider the category of finite dimensinal $k$-representations of $G$.

GroupRepresentationCategory{T,G} <: RepresentationCategory{T}

Build it with the constructor

JuCat.RepresentationCategory — Method

RepresentationCategory(G::GAPGroup, F::Field)

Category of finite dimensonal group representations of \G.

A group representation is defined by a group homomorphism from $G$ into a finite dimensional vector space $k^n$. These objects are of type

GroupRepresentation{T,G} <: Representation

They are constructed in one of two ways, either by images of generators or by a function

JuCat.Representation — Method

Representation(G::GAPGroup, pre_img::Vector, img::Vector)

Group representation defined by the images of generators of G.

JuCat.Representation — Method

Representation(G::GAPGroup, m::Function)

Group representation defined by m:G -> Mat_n.

where in both cases the images are required to be fitting MatrixElem objects.

Since group reprsentation categories are tensor categories, we again have methods for the important operations

JuCat.Morphism — Method

Morphism(ρ::GroupRepresentation{T,G}, τ::GroupRepresentation{T,G}, m::MatElem{T}; check = true) where {T,G}

Morphism between representations defined by $m$. If check == false equivariancy will not be checked.

AbstractAlgebra.isisomorphic — Method

isisomorphic(σ::GroupRepresentation{T,G}, τ::GroupRepresentation{T,G}) where {T,G}

Check whether σ and τ are isomorphic. If true return the isomorphism.

Base.one — Method

one(Rep::GroupRepresentationCategory{T,G}) where {T,G}

Return the trivial representation.

Base.parent — Method

parent(ρ::GroupRepresentation{T,G}) where {T,G}

Return the parent representation category of ρ.

Base.zero — Method

zero(Rep::GroupRepresentationCategory{T,G}) where {T,G}

Return the zero reprensentation.

Hecke.id — Method

id(ρ::GroupRepresentation{T,G}) where {T,G}

Return the identity on ρ.

Hecke.tensor_product — Method

tensor_product(f::GroupRepresentationMorphism, g::GroupRepresentationMorphism)

Return the tensor product of morphisms of representations.

Hecke.tensor_product — Method

tensor_product(ρ::GroupRepresentation{T}, τ::GroupRepresentation{T}) where T

Return the tensor product of representations.

JuCat.Hom — Method

Hom(σ::GroupRepresentation{T,G}, τ::GroupRepresentation{T,G}) where {T,G}

Return the hom-space of the representations as a vector space.

JuCat.decompose — Method

decompose(σ::GroupRepresentation)

Decompose the representation into a direct sum of simple objects. Return a list of tuples with simple objects and multiplicities.

JuCat.dsum — Method

dsum(f::GroupRepresentationMorphism, g::GroupRepresentationMorphism)

Direct sum of morphisms of representations.

JuCat.dsum — Method

dsum(ρ::GroupRepresentation{T,G}, τ::GroupRepresentation{T,G}, morphisms::Bool = false) where {T,G}

Return the direct sum of representations. If morphisms is set true inclusion and projection morphisms are also returned.

JuCat.simples — Method

simples(Rep::GroupRepresentationCategory{T,G}) where {T,G}

Return a list of the simple objects in Rep.

Oscar.issemisimple — Method

issemisimple(C::GroupRepresentationCategory)

Return true if C is semisimple else false.