Vector Space Categories

Vector spaces in JuCat are of the abstract type

JuCat.VectorSpaceObject — Type

VectorSpaceObject{T}

An object in the category of finite dimensional vector spaces.

They always need the following fields:

basis::Vector{Any}
parent::Category

The Category of Finite Dimensional VectorSpaces

The simplest example to provide are the finite dimensional vector spaces over a field.

The basic methods to generate this are

JuCat.VectorSpaces — Type

VectorSpaces{T}(K::S) where T <: FieldElem

The category of finite dimensional vector spaces over K.

The following methods are provided:

JuCat.Cocycle — Method

Cocylce(G::GAPGroup, m::Dict{NTuple{N,G}, T})

Return a N-cocylce of G. By now the condition is not checked.

JuCat.GradedVectorSpaces — Type

GradedVectorSpaces{T}(K::S) where T <: FieldElem

The category of finite dimensional vector spaces over K.

JuCat.Morphism — Method

Morphism(X::VectorSpaceObject, Y::VectorSpaceObject, m::MatElem)

Return a morphism in the category of vector spaces defined by m.

JuCat.VectorSpaceMorphism — Type

VectorSpaceMorphism{T}()

A morphism in the category of finite dimensional vector spaces.

JuCat.VectorSpaceObject — Method

VectorSpaceObject(Vec::VectorSpaces, n::Int64)
VectorSpaceObject(K::Field, n::Int)
VectorSpaceObject(Vec::VectorSpaces, basis::Vector)
VectorSpaceObject(K::Field, basis::Vector)

The n-dimensional vector space with basis v1,..,vn (or other specified basis)

AbstractAlgebra.base_ring — Method

base_ring(C::Category)

Return the base ring kof the k-linear category C.

AbstractAlgebra.base_ring — Method

base_ring(X::Object)

Return the base ring k of the k-linear parent category of X.

Base.:^ — Method

^(X::Object, n::Integer)

Return the n-fold product object X^n.

Base.parent — Method

parent(X::Object)

Return the parent category of the object X.

Hecke.:⊗ — Method

⊗(X::Object...)

Return the tensor product object.

Hecke.:⊗ — Method

⊗(f::Morphism, g::Morphism)

Return the tensor product morphism of fand g.

JuCat.:× — Method

×(X::Object...)

Return the product Object and an array containing the projection morphisms.

JuCat.dsum — Method

dsum(X::VectorSpaceObject{T,S}...) where {T,S <: FreeModule}

Direct sum space of X... together with the embedding morphisms.

JuCat.∐ — Method

∐(X::Object...)

Return the coproduct Object and an array containing the injection morphisms.

The Category of Graded Vector Spaces