Vector Space Categories

Vector spaces in JuCat are of the abstract type

VectorSpaceObject{T} <: Object

All objects with vector space structure like hom-spaces are and should be implemented as a subtype of this type. They always need the following fields:

basis::Vector{Any}
parent::Category

The Category of Finite Dimensional VectorSpaces

The simplest example to provide are the finite dimensional vector spaces over a field. This category has type

VectorSpaces{T} <: TensorCategory{T}

and can be constructed like so:

using Oscar,JuCat # hide
F = FiniteField(5,2)
Vec = VectorSpaces(F)

Objects of this category are of the type

VSObject{T} <: VectorSpaceObject{T}

Every vector space object is defined by a basis and a base field provided by the parent category.

JuCat.VectorSpaceObject — Type

VectorSpaceObject{T}

An object in the category of finite dimensional vector spaces.

JuCat.VectorSpaceObject — Method

VectorSpaceObject(Vec::VectorSpaces, n::Int64)
VectorSpaceObject(K::Field, n::Int)
VectorSpaceObject(Vec::VectorSpaces, basis::Vector)
VectorSpaceObject(K::Field, basis::Vector)

The n-dimensional vector space with basis v1,..,vn (or other specified basis)

The Category of Graded Vector Spaces

Very similar we have the category of (twisted) $G$-graded vector spaces for a finite group $G$. We have the type

Missing docstring.

Missing docstring for GradedVectorSpaces{T,G} <: VectorSpaces{T,G}. Check Documenter's build log for details.

and they are constructed in straightforward manner

julia> G = symmetric_group(6)ERROR: UndefVarError: symmetric_group not defined
julia> F,a = FiniteField(2,3)ERROR: UndefVarError: FiniteField not defined
julia> VecG = GradedVectorSpaces(G,F)ERROR: UndefVarError: G not defined

To add a nontrivial associator (twist) construct a Cocycle{3} object coding a 3-cocycle of the group $G$. By now no checking of this condition happens.

julia> C = #TODO CoCycle(G, )
       VecG = #VectorSpaces(G,QQ,C)ERROR: syntax: incomplete: premature end of input

Functionality

(Graded) vector spaces form a semisimple tensor category. Thus the methods for direct sums, standard tensor products, one and zero object are all implemented.

AbstractAlgebra.base_ring — Method

base_ring(C::Category)

Return the base ring kof the k-linear category C.

AbstractAlgebra.base_ring — Method

base_ring(X::Object)

Return the base ring k of the k-linear parent category of X.

Base.:^ — Method

^(X::Object, n::Integer)

Return the n-fold product object X^n.

Base.parent — Method

parent(X::Object)

Return the parent category of the object X.

Hecke.:⊗ — Method

⊗(X::Object...)

Return the tensor product object.

Hecke.:⊗ — Method

⊗(f::Morphism, g::Morphism)

Return the tensor product morphism of fand g.

JuCat.:× — Method

×(X::Object...)

Return the product Object and an array containing the projection morphisms.

JuCat.dsum — Method

dsum(X::VectorSpaceObject{T,S}...) where {T,S <: FreeModule}

Direct sum space of X... together with the embedding morphisms.

JuCat.∐ — Method

∐(X::Object...)

Return the coproduct Object and an array containing the injection morphisms.