Vector Space Categories

Vector spaces in TensorCategories are of the abstract type

abstract type VectorSpaceCategoryObject <: CategoryObject end

All objects with vector space structure like hom-spaces are and should be implemented as a subtype of this type. They always need the following fields:

basis::Vector
parent::Category

Finite Dimensional VectorSpaces

The simplest example to provide are the finite dimensional vector spaces over a field. This category has type

VectorSpaces <: TensorCategory

and can be constructed like so:

F,a = FiniteField(5,2)
Vec = VectorSpaces()

CategoryObjects of this category are of the type

VSCategoryObject <: VectorSpaceCategoryObject

Every vector space object is defined by a basis and a base field provided by the parent category.

TensorCategories.VectorSpaceCategoryObject — Type

VectorSpaceCategoryObject

An object in the category of finite dimensional vector spaces.

TensorCategories.VectorSpaceCategoryObject — Method

VectorSpaceCategoryObject(Vec::VectorSpaces, n::Int64)
VectorSpaceCategoryObject(K::Field, n::Int)
VectorSpaceCategoryObject(Vec::VectorSpaces, basis::Vector)
VectorSpaceCategoryObject(K::Field, basis::Vector)

The n-dimensional vector space with basis v1,..,vn (or other specified basis)

Morphisms in this Category are defined only by matrices of matching dimensions. They are typed as

VSCategoryMorphism <: CategoryMorphism

and constructed giving a domain, codomain and matrix element.

CategoryMorphism(::VectorSpaceCategoryObject, ::VectorSpaceCategoryObject, ::MatElem)

Graded Vector Spaces

Very similar we have the category of finite dimensional (twisted) $G$-graded vector spaces for a finite group $G$. We have the type

GradedVectorSpaces <: VectorSpaces

and they are constructed in straightforward manner

G = symmetric_group(6)
VecG = GradedVectorSpaces(G)

Category of G-graded vector spaces over Field of algebraic numbers where G is Sym( [ 1 .. 6 ] )

To add a non-trivial associator (twist) there is another constructor.

TensorCategories.TwistedGradedVectorSpaces — Function

TwistedGradedVectorSpaces(G::GAPGroup, i::Int)

Construct the category of twisted graded vectorspaces with the i-th 3-cocycle.

Graded vector spaces decompose into direct sums of vector spaces for each element in $G$.

GVSCategoryObject <: VectorSpaceCategoryObject

G = symmetric_group(5)
g,s = gens(G)
V1 = VectorSpaceCategoryObject(QQ,5)
V2 = VectorSpaceCategoryObject(QQ, [:v, :w])
W = VectorSpaceCategoryObject(g => V1, s => V2, g*s => V1⊗V2)

Graded vector space of dimension 17 with grading
PermGroupElem[(1,2,3,4,5), (1,2,3,4,5), (1,2,3,4,5), (1,2,3,4,5), (1,2,3,4,5), (1,2), (1,2), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5)]

Morphisms are implemented analogously by pairs of group elements and vector space objects.

GVSCategoryMorphism <: CategoryMorphism

The constructor is given by

Missing docstring.

Missing docstring for Morphism(::GVSCategoryObject, ::GVSCategoryObject,::MatElem) where {G, S <: VectorSpaceCategoryMorphism}. Check Documenter's build log for details.

Functionality

(Graded) vector spaces form a fusion category. Thus the methods for direct sums, tensor products, dual, one and zero object are all implemented.

AbstractAlgebra.Generic.dim — Method

dim(V::VectorSpaceCategoryObject) = length(V.basis)

Return the vector space dimension of $V$.

AbstractAlgebra.direct_sum — Method

function direct_sum(V::GVSCategoryObject, W::GVSCategoryObject)

Return the direct sum object $V⊕W$.

AbstractAlgebra.direct_sum — Method

direct_sum(f::VectorSpaceCategoryMorphism{T},g::VectorSpaceCategoryMorphism{T}) where T

Return the direct sum of morphisms of vector spaces.

AbstractAlgebra.direct_sum — Method

direct_sum(X::VectorSpaceCategoryObject{T}, Y::VectorSpaceCategoryObject{T}) where {T}

Direct sum of vector spaces together with the embedding morphisms.

AbstractAlgebra.is_isomorphic — Method

function is_isomorphic(V::GVSCategoryObject, W::GVSCategoryObject)

Check whether $V$and $W$are isomorphic as $G$-graded vector spaces and return an isomorphism in the positive case.

AbstractAlgebra.is_isomorphic — Method

is_isomorphic(V::VSCategoryObject, W::VSCategoryObject)

Check whether $V$ and $W$are isomorphic. Return the isomorphisms if existent.

AbstractAlgebra.kernel — Method

function kernel(f::GVSCategoryMorphism)

Return the graded vector space kernel of $f$.

Base.one — Method

function one(C::GradedVectorSpaces)

Return $k$ as the one dimensional graded vector space.

Base.one — Method

one(Vec::VectorSpaces) = VectorSpaceCategoryObject(base_ring(Vec),1)

Return the one-dimensional vector space.

Base.zero — Method

function zero(C::GradedVectorSpaces)

Return the zero diemsnional graded vector space.

Base.zero — Method

zero(Vec::VectorSpaces) = VectorSpaceCategoryObject(base_ring(Vec), 0)

Return the zero-dimensional vector space.

Hecke.cokernel — Method

function cokernel(f::GVSCategoryMorphism)

Return the graded vector space cokernel of $f$.

Hecke.decompose — Method

function decompose(V::GVSCategoryObject)

Return a vector with the simple objects together with their multiplicities $[V:Xi]$.

Hecke.dual — Method

function dual(V::GVSCategoryObject)

Return the graded dual vector space of $V$.

Hecke.id — Method

id(X::VectorSpaceCategoryObject{T}) where T

Return the identity on the vector space $X$.

Hecke.tensor_product — Method

function tensor_product(V::GVSCategoryObject, W::GVSCategoryObject)

Return the tensor product $V⊗W$.

Hecke.tensor_product — Method

tensor_product(f::VectorSpaceCategoryMorphism, g::VectorSpaceCategoryMorphism)

Return the tensor product of vector space morphisms.

Hecke.tensor_product — Method

tensor_product(X::VectorSpaceCategoryObject{T}, Y::VectorSpaceCategoryObject{T}) where {T,S1,S2}

Return the tensor product of vector spaces.

TensorCategories.Hom — Method

function Hom(V::GVSCategoryObject, W::GVSCategoryObject)

Return the space of morphisms between graded vector spaces $V$ and $W$.

TensorCategories.Hom — Method

Hom(X::VectorSpaceCategoryObject, Y::VectorSpaceCategoryObject)

Return the Hom($X,Y$`) as a vector space.

TensorCategories.Morphism — Method

function Morphism(V::GVSCategoryObject, Y::GVSCategoryObject, m::MatElem)

Return the morphism $V → W$defined by $m$.

TensorCategories.Morphism — Method

Morphism(m::MatElem)

Vector space morphisms defined by m.

TensorCategories.Morphism — Method

Morphism(X::VectorSpaceCategoryObject, Y::VectorSpaceCategoryObject, m::MatElem)

Return a morphism in the category of vector spaces defined by m.

TensorCategories.associator — Method

function associator(U::GVSCategoryObject, V::GVSCategoryObject, W::GVSCategoryObject)

return the associator isomorphism $(U⊗V)⊗W → U⊗(V⊗W)$.

TensorCategories.associator — Method

associator(X::VectorSpaceCategoryObject, Y::VectorSpaceCategoryObject, Z::VectorSpaceCategoryObject)

Return the associator isomorphism a::(X⊗Y)⊗Z -> X⊗(Y⊗Z).

TensorCategories.ev — Method

function ev(V::GVSCategoryObject)

Return the evaluation map $V*⊗V → 𝟙$.

TensorCategories.simples — Method

function simples(C::GradedVectorSpaces)

Return a vector containing the simple objects of $C$.