Vector Space Categories

Vector spaces in TensorCategories are of the abstract type

abstract type VectorSpaceObject{T} <: Object end

All objects with vector space structure like hom-spaces are and should be implemented as a subtype of this type. They always need the following fields:

basis::Vector{Any}
parent::Category

Finite Dimensional VectorSpaces

The simplest example to provide are the finite dimensional vector spaces over a field. This category has type

VectorSpaces <: TensorCategory

and can be constructed like so:

F,a = FiniteField(5,2)
Vec = VectorSpaces(F)

Category of finite dimensional VectorSpaces over Finite field of degree 2 over F_5

Objects of this category are of the type

VSObject <: VectorSpaceObject

Every vector space object is defined by a basis and a base field provided by the parent category.

TensorCategories.VectorSpaceObject — Type

VectorSpaceObject{T}

An object in the category of finite dimensional vector spaces.

TensorCategories.VectorSpaceObject — Method

VectorSpaceObject(Vec::VectorSpaces, n::Int64)
VectorSpaceObject(K::Field, n::Int)
VectorSpaceObject(Vec::VectorSpaces, basis::Vector)
VectorSpaceObject(K::Field, basis::Vector)

The n-dimensional vector space with basis v1,..,vn (or other specified basis)

Morphisms in this Category are defined only by matrices of matching dimensions. They are typed as

VSMorphism <: Morphism

and constructed giving a domain, codomain and matrix element.

TensorCategories.Morphism — Method

Morphism(X::VectorSpaceObject, Y::VectorSpaceObject, m::MatElem)

Return a morphism in the category of vector spaces defined by m.

Graded Vector Spaces

Very similar we have the category of finite dimensional (twisted) $G$-graded vector spaces for a finite group $G$. We have the type

GradedVectorSpaces <: VectorSpaces

and they are constructed in straightforward manner

G = symmetric_group(6)
F,a = FiniteField(2,3)
VecG = GradedVectorSpaces(F,G)

Category of G-graded vector spaces over Finite field of degree 3 over F_2 where G is Sym( [ 1 .. 6 ] )

To add a nontrivial associator (twist) construct a Cocycle{3} object coding a 3-cocycle of the group $G$. By now no checking of this condition happens.

C = #TODO CoCycle(G, )
VecG = #VectorSpaces(G,QQ,C)

Graded vector spaces decompose into direct sums of vector spaces for each element in $G$.

GVSObject <: VectorSpaceObject

G = symmetric_group(5)
g,s = gens(G)
V1 = VectorSpaceObject(QQ,5)
V2 = VectorSpaceObject(QQ, [:v, :w])
W = VectorSpaceObject(g => V1, s => V2, g*s => V1⊗V2)

Graded vector space of dimension 17 with grading
PermGroupElem[(1,2,3,4,5), (1,2,3,4,5), (1,2,3,4,5), (1,2,3,4,5), (1,2,3,4,5), (1,2), (1,2), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5), (2,3,4,5)]

Morphisms are implemented analogously by pairs of group elements and vectorspace objects.

GVSMorphism <: Morphism

The constructors are

Missing docstring.

Missing docstring for Morphism(::GVSObject,::GVSObject,::Dict{G,S}) where {S<:VectorSpaceMorphism,G}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Morphism(::GVSObject, ::GVSObject,::Pair{G,S}...) where {G, S <: VectorSpaceMorphism. Check Documenter's build log for details.

Functionality

(Graded) vector spaces form a semisimple tensor category. Thus the methods for direct sums, standard tensor products, one and zero object are all implemented.

AbstractAlgebra.Generic.dim — Method

dim(V::VectorSpaceObject) = length(V.basis)

Return the vector space dimension of $V$.

AbstractAlgebra.isisomorphic — Method

isisomorphic(V::VSObject, W::VSObject)

Check whether $V$ and $W$are isomorphic. Return the isomorphisms if existent.

Base.one — Method

one(Vec::VectorSpaces) = VectorSpaceObject(base_ring(Vec),1)

Return the one-dimensional vector space.

Base.zero — Method

zero(Vec::VectorSpaces) = VectorSpaceObject(base_ring(Vec), 0)

Return the zero-dimensional vector space.

Hecke.id — Method

id(X::VectorSpaceObject{T}) where T

Return the identity on the vector space $X$.

Hecke.tensor_product — Method

tensor_product(f::VectorSpaceMorphism, g::VectorSpaceMorphism)

Return the tensor product of vector space morphisms.

Hecke.tensor_product — Method

tensor_product(X::VectorSpaceObject{T}, Y::VectorSpaceObject{T}) where {T,S1,S2}

Return the tensor product of vector spaces.

TensorCategories.Hom — Method

Hom(X::VectorSpaceObject, Y::VectorSpaceObject)

Return the Hom($X,Y$`) as a vector space.

TensorCategories.associator — Method

associator(X::VectorSpaceObject, Y::VectorSpaceObject, Z::VectorSpaceObject)

Return the associator isomorphism a::(X⊗Y)⊗Z -> X⊗(Y⊗Z).

TensorCategories.dsum — Function

dsum(X::VectorSpaceObject{T}, Y::VectorSpaceObject{T}, morphisms = false) where {T}

Direct sum of vector spaces together with the embedding morphisms if morphisms = true.

TensorCategories.dsum — Method

dsum(f::VectorSpaceMorphism{T},g::VectorSpaceMorphism{T}) where T

Return the direct sum of morphisms of vector spaces.