Vector Space Categories

Vector spaces in JuCat are of the abstract type

abstract type VectorSpaceObject{T} <: Object end

All objects with vector space structure like hom-spaces are and should be implemented as a subtype of this type. They always need the following fields:

basis::Vector{Any}
parent::Category

Finite Dimensional VectorSpaces

The simplest example to provide are the finite dimensional vector spaces over a field. This category has type

VectorSpaces{T} <: TensorCategory{T}

and can be constructed like so:

F,a = FiniteField(5,2)
Vec = VectorSpaces(F)

Category of finite dimensional VectorSpaces over Finite field of degree 2 over F_5

Objects of this category are of the type

VSObject{T} <: VectorSpaceObject{T}

Every vector space object is defined by a basis and a base field provided by the parent category.

JuCat.VectorSpaceObject — Type

VectorSpaceObject{T}

An object in the category of finite dimensional vector spaces.

JuCat.VectorSpaceObject — Method

VectorSpaceObject(Vec::VectorSpaces, n::Int64)
VectorSpaceObject(K::Field, n::Int)
VectorSpaceObject(Vec::VectorSpaces, basis::Vector)
VectorSpaceObject(K::Field, basis::Vector)

The n-dimensional vector space with basis v1,..,vn (or other specified basis)

Morphisms in this Category are defined only by matrices of matching dimensions. They are typed as

VSMorphism{T} <: Morphism

and constructed giving a domain, codomain and matrix element.

JuCat.Morphism — Method

Morphism(X::VectorSpaceObject, Y::VectorSpaceObject, m::MatElem)

Return a morphism in the category of vector spaces defined by m.

Graded Vector Spaces

Very similar we have the category of finite dimensional (twisted) $G$-graded vector spaces for a finite group $G$. We have the type

GradedVectorSpaces{T,G} <: VectorSpaces{T,G}

and they are constructed in straightforward manner

G = symmetric_group(6)
F,a = FiniteField(2,3)
VecG = GradedVectorSpaces(F,G)

Category of finite dimensional graded vector spaces over Finite field of degree 3 over F_2
with simple objects in Sym( [ 1 .. 6 ] ).

To add a nontrivial associator (twist) construct a Cocycle{3} object coding a 3-cocycle of the group $G$. By now no checking of this condition happens.

C = #TODO CoCycle(G, )
VecG = #VectorSpaces(G,QQ,C)

Graded vector spaces decompose into direct sums of vector spaces for each element in $G$.

GVSObject{T,G} <: VectorSpaceObject{T}

G = symmetric_group(5)
g,s = gens(G)
V1 = VectorSpaceObject(QQ,5)
V2 = VectorSpaceObject(QQ, [:v, :w])
W = VectorSpaceObject(g => V1, s => V2, g*s => V1⊗V2)

Graded Vector Space of dimension 17 over Rational Field. Direct sum of 
   (1,2,3,4,5) ⋅ Vector space of dimension 5 over Rational Field.
   (1,2) ⋅ Vector space of dimension 2 over Rational Field.
   (2,3,4,5) ⋅ Vector space of dimension 10 over Rational Field.

Morphisms are implemented analogously by pairs of group elements and vectorspace objects.

GVSMorphism{T,G} <: Morphism

The constructors are

JuCat.Morphism — Method

Morphism(D::GVSObject, C::GVSObject, m::Dict)

Morphism between graded vector spaces defined by pairs of elements and morphisms g => (Dg -> Cg).

JuCat.Morphism — Method

Morphism(D::GVSObject, C::GVSObject, m::Pair...)

Morphism between graded vector spaces defined by pairs of elements and morphisms g => (Dg -> Cg).

Functionality

(Graded) vector spaces form a semisimple tensor category. Thus the methods for direct sums, standard tensor products, one and zero object are all implemented.

JuCat.dsum — Method

dsum(X::VectorSpaceObject{T,S}...) where {T,S <: FreeModule}

Direct sum space of X... together with the embedding morphisms.